[WEB-15] refactored the arrows component
