+++ src/components/Testimonials/Testimonials.tsx
@@ -5,7 +5,7 @@ import "slick-carousel/slick/slick-theme.css";
 import { TESTIMONIALS } from "@/prototypes/testimonials";
 import React, { MutableRefObject, useRef } from "react";
 import Slider from "react-slick";
-import { NextArrow, PreviousArrow, Testimony } from ".";
+import { SliderArrow, Testimony } from ".";
 
 const Testimonials = () => {
   let sliderRef = useRef<Slider | null>(null);
@@ -60,8 +60,8 @@ const Testimonials = () => {
         </Slider>
       </div>
       <div className="flex-center gap-5">
-        <PreviousArrow onClick={previous} />
-        <NextArrow onClick={next} />
+        <SliderArrow onClick={previous} direction="left" />
+        <SliderArrow onClick={next} direction="right" />
       </div>
     </div>
   );
