+++ src/components/Testimonials/index.ts
@@ -1,11 +1,5 @@
 import Testimonails from "./Testimonials";
 import Testimony from "./Testimony";
 import TestimonialVideoModal from "./TestimonialVideoModal";
-import { PreviousArrow, NextArrow } from "./PrevNextArrows";
-export {
-  Testimonails,
-  Testimony,
-  TestimonialVideoModal,
-  PreviousArrow,
-  NextArrow,
-};
+import { SliderArrow } from "./SliderArrow";
+export { Testimonails, Testimony, TestimonialVideoModal, SliderArrow };
